requirements.txt
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
openai>=1.6.1,<2.0.0
chromadb==0.4.15
requests==2.31.0
httpx==0.25.2
pyyaml==6.0.1
python-dotenv==1.0.0
streamlit==1.28.1
pydantic==2.4.2
typing-extensions==4.8.0
